% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_responses.R
\name{scale_responses}
\alias{scale_responses}
\title{Scale survey responses by min-max}
\usage{
scale_responses(x)
}
\arguments{
\item{x}{A numeric vector of responses}
}
\value{
A numeric vector scaled to 0-1 range.
}
\description{
Convenience function to scale Likert-type responses to the 0-1 range.
}
\examples{
x <- 1:5
scale_responses(x)
}
