% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_keywords.R
\name{split_keywords}
\alias{split_keywords}
\title{Split comma-separated survey keywords}
\usage{
split_keywords(x, sep = ",")
}
\arguments{
\item{x}{A character vector with one element.}

\item{sep}{A single character to split on, default is ",".}
}
\value{
A character vector of trimmed keywords.
}
\description{
Useful for open-ended responses such as "anxiety,depression".
}
\examples{
split_keywords("anxiety, depression, stress")
split_keywords(c("happiness, satisfaction", "fatigue, lethargy"))
}
